+++ zit/GeneratedVersion.cs
@@ -2,5 +2,5 @@ namespace Zit;
 
 public static class GeneratedVersion
 {
-    public const string Value = "0.1.2";
+    public const string Value = "0.1.3";
 }